+++ wp-content/plugins/newspack-popups/includes/class-newspack-popups-inserter.php
@@ -7,6 +7,8 @@
 
 defined( 'ABSPATH' ) || exit;
 
+require_once dirname( __FILE__ ) . '/../api/segmentation/class-segmentation.php';
+
 /**
  * API endpoints
  */
@@ -53,71 +55,89 @@ final class Newspack_Popups_Inserter {
 			return [ Newspack_Popups_Model::retrieve_preview_popup( Newspack_Popups::previewed_popup_id() ) ];
 		}
 
-		// 1. Get all inline popups.
-		$popups_to_maybe_display = Newspack_Popups_Model::retrieve_inline_popups();
-
-		// 2. Get the overlay popup/s. There can be only one displayed, unless in test mode.
-
-		// Any overlay test popups, if the user is logged in.
-		if ( is_user_logged_in() ) {
-			$popups_to_maybe_display = array_merge(
-				$popups_to_maybe_display,
-				Newspack_Popups_Model::retrieve_overlay_test_popups()
-			);
+		// Campaigns disabled for this page.
+		if ( self::assess_has_disabled_popups() ) {
+			return [];
 		}
 
-		// Check if there's an overlay popup with matching category.
-		$category_overlay_popup = Newspack_Popups_Model::retrieve_category_overlay_popup();
-		if ( $category_overlay_popup && self::should_display( $category_overlay_popup ) ) {
-			array_push(
-				$popups_to_maybe_display,
-				$category_overlay_popup
-			);
+		$view_as_spec             = Segmentation::parse_view_as( Newspack_Popups_View_As::viewing_as_spec() );
+		$view_as_spec_groups      = isset( $view_as_spec['groups'] ) ? $view_as_spec['groups'] : false;
+		$view_as_spec_all         = ! empty( $view_as_spec['all'] ) ? true : false;
+		$view_as_spec_segment     = isset( $view_as_spec['segment'] ) ? $view_as_spec['segment'] : false;
+		$view_as_spec_unpublished = isset( $view_as_spec['show_unpublished'] ) && 'true' === $view_as_spec['show_unpublished'] ? true : false;
+
+		if ( $view_as_spec_groups ) {
+			// If previewing specific groups.
+			$popups_to_maybe_display = Newspack_Popups_Model::retrieve_group_popups( explode( ',', $view_as_spec['groups'] ), $view_as_spec_unpublished );
+		} elseif ( $view_as_spec_all || $view_as_spec_segment ) {
+			// If previewing and no groups are specified, but 'all' or a segment is specified, retrieve all campaigns.
+			$popups_to_maybe_display = Newspack_Popups_Model::retrieve_popups( $view_as_spec_unpublished );
 		} else {
-			// If there's no category-matching popup, get the sitewide pop-up.
-			$sitewide_default = get_option( Newspack_Popups::NEWSPACK_POPUPS_SITEWIDE_DEFAULT, null );
-			if ( $sitewide_default ) {
-				$found_popup = Newspack_Popups_Model::retrieve_popup_by_id( $sitewide_default );
-				if (
-					$found_popup &&
-					// Prevent inline sitewide default from being added - all inline popups are there.
-					'inline' !== $found_popup['options']['placement']
-				) {
-					array_push(
-						$popups_to_maybe_display,
-						$found_popup
-					);
+			// Retrieve campaigns for front-end display.
+
+			// 1. Get all inline popups.
+			$popups_to_maybe_display = Newspack_Popups_Model::retrieve_inline_popups();
+
+			// Check if there's an overlay popup with matching category.
+			$category_overlay_popup = Newspack_Popups_Model::retrieve_category_overlay_popup();
+			if ( $category_overlay_popup && self::should_display( $category_overlay_popup ) ) {
+				array_push(
+					$popups_to_maybe_display,
+					$category_overlay_popup
+				);
+			} else {
+				// If there's no category-matching popup, get the sitewide pop-up.
+				$sitewide_default = get_option( Newspack_Popups::NEWSPACK_POPUPS_SITEWIDE_DEFAULT, null );
+				if ( $sitewide_default ) {
+					$found_popup = Newspack_Popups_Model::retrieve_popup_by_id( $sitewide_default );
+					if (
+						$found_popup &&
+						// Prevent non-overlay sitewide default from being added.
+						Newspack_Popups_Model::is_overlay( $found_popup )
+					) {
+						array_push(
+							$popups_to_maybe_display,
+							$found_popup
+						);
+					}
 				}
 			}
 		}
 
 		// Allow only one overlay campaign.
-		$has_overlay                     = false;
-		$popups_to_maybe_display_deduped = array_filter(
-			$popups_to_maybe_display,
-			function ( $campaign ) use ( &$has_overlay ) {
-				if ( 'inline' !== $campaign['options']['placement'] ) {
-					if ( $has_overlay ) {
-						return false;
-					} else {
-						$has_overlay = true;
-						return true;
+		if ( empty( $view_as_spec ) ) {
+			$has_overlay                     = false;
+			$popups_to_maybe_display_deduped = array_filter(
+				$popups_to_maybe_display,
+				function ( $campaign ) use ( &$has_overlay ) {
+					if ( Newspack_Popups_Model::is_overlay( $campaign ) ) {
+						if ( $has_overlay ) {
+							return false;
+						} else {
+							$has_overlay = true;
+							return true;
+						}
 					}
+					return true;
 				}
-				return true;
+			);
+		} else {
+			// If previewing, allow all matching overlay campaigns to be displayed.
+			$popups_to_maybe_display_deduped = $popups_to_maybe_display;
+		}
+
+		// Remove manual placement campaigns.
+		$popups_to_maybe_display_deduped = array_filter(
+			$popups_to_maybe_display_deduped,
+			function( $campaign ) {
+				return 'manual' !== $campaign['options']['frequency'];
 			}
 		);
 
-		$popups_to_display = array_filter(
+		return array_filter(
 			$popups_to_maybe_display_deduped,
 			[ __CLASS__, 'should_display' ]
 		);
-
-		if ( ! empty( $popups_to_display ) ) {
-			return $popups_to_display;
-		}
-
-		return [];
 	}
 
 	/**
@@ -129,9 +149,18 @@ final class Newspack_Popups_Inserter {
 		add_action( 'after_header', [ $this, 'insert_popups_after_header' ] ); // This is a Newspack theme hook. When used with other themes, popups won't be inserted on archive pages.
 		add_action( 'wp_head', [ $this, 'insert_popups_amp_access' ] );
 		add_action( 'wp_head', [ $this, 'register_amp_scripts' ] );
+		add_action( 'before_header', [ $this, 'insert_before_header' ] );
 
 		// Always enqueue scripts, since this plugin's scripts are handling pageview sending via GTAG.
 		add_action( 'wp_enqueue_scripts', [ $this, 'enqueue_scripts' ] );
+
+		add_filter(
+			'widget_update_callback',
+			[ $this, 'save_widgets_shortcoded_popup_ids' ],
+			1,
+			4
+		);
+		add_action( 'delete_widget', [ $this, 'remove_widgets_shortcoded_popup_ids' ] );
 
 		add_filter(
 			'newspack_newsletters_assess_has_disabled_popups',
@@ -173,9 +202,8 @@ final class Newspack_Popups_Inserter {
 	 * Process popups and insert into post and page content if needed.
 	 *
 	 * @param string $content The content of the post.
-	 * @param bool   $enqueue_assets Whether assets should be enqueued.
 	 */
-	public static function insert_popups_in_content( $content = '', $enqueue_assets = true ) {
+	public static function insert_popups_in_content( $content = '' ) {
 		// Avoid duplicate execution.
 		if ( true === self::$the_content_has_rendered ) {
 			return $content;
@@ -201,13 +229,8 @@ final class Newspack_Popups_Inserter {
 			return $content;
 		}
 
-		// Campaigns disabled for this page.
-		if ( self::assess_has_disabled_popups() ) {
-			return $content;
-		}
-
 		// If the current post is a Campaign, ignore.
-		if ( Newspack_Popups::NEWSPACK_PLUGINS_CPT == get_post_type() ) {
+		if ( Newspack_Popups::NEWSPACK_POPUPS_CPT == get_post_type() ) {
 			return $content;
 		}
 
@@ -217,7 +240,14 @@ final class Newspack_Popups_Inserter {
 			return $content;
 		}
 
-		$popups = self::popups_for_post();
+		// If any popups are inserted using a shortcode, skip them.
+		$shortcoded_popups_ids = self::get_shortcoded_popups_ids( get_the_content() );
+		$popups                = array_filter(
+			self::popups_for_post(),
+			function ( $popup ) use ( $shortcoded_popups_ids ) {
+				return ! in_array( $popup['id'], $shortcoded_popups_ids ) && Newspack_Popups_Model::should_be_inserted_in_page_content( $popup );
+			}
+		);
 
 		if ( empty( $popups ) ) {
 			return $content;
@@ -239,14 +269,19 @@ final class Newspack_Popups_Inserter {
 		$inline_popups  = [];
 		$overlay_popups = [];
 		foreach ( $popups as $popup ) {
-			if ( 'inline' === $popup['options']['placement'] ) {
+			if ( Newspack_Popups_Model::is_inline( $popup ) ) {
 				$percentage                = intval( $popup['options']['trigger_scroll_progress'] ) / 100;
 				$popup['precise_position'] = $total_length * $percentage;
 				$popup['is_inserted']      = false;
 				$inline_popups[]           = $popup;
-			} else {
+			} elseif ( Newspack_Popups_Model::is_overlay( $popup ) ) {
 				$overlay_popups[] = $popup;
 			}
+		}
+
+		// Return early if there are no popups to insert. This can happen if e.g. the only popup is an above header one.
+		if ( empty( $inline_popups ) && empty( $overlay_popups ) ) {
+			return $content;
 		}
 
 		// 2. Iterate overall blocks and insert inline campaigns.
@@ -291,13 +326,19 @@ final class Newspack_Popups_Inserter {
 		if ( is_singular() ) {
 			return;
 		}
+		$popups = array_filter( self::popups_for_post(), [ 'Newspack_Popups_Model', 'should_be_inserted_in_page_content' ] );
+		foreach ( $popups as $popup ) {
+			echo Newspack_Popups_Model::generate_popup( $popup ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		}
+	}
 
-		$popups = self::popups_for_post();
-
-		if ( ! empty( $popups ) ) {
-			foreach ( $popups as $popup ) {
-				echo Newspack_Popups_Model::generate_popup( $popup ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-			}
+	/**
+	 * Insert popups markup before header.
+	 */
+	public static function insert_before_header() {
+		$before_header_popups = array_filter( self::popups_for_post(), [ 'Newspack_Popups_Model', 'should_be_inserted_above_page_header' ] );
+		foreach ( $before_header_popups as $popup ) {
+			echo Newspack_Popups_Model::generate_popup( $popup ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 		}
 	}
 
@@ -332,6 +373,8 @@ final class Newspack_Popups_Inserter {
 
 	/**
 	 * The popup shortcode function.
+	 * Primarily, the shortcode is inserted by the plugin, but it may also be inserted manually to
+	 * display a specific campaign anywhere on the site.
 	 *
 	 * @param array $atts Shortcode attributes.
 	 * @return HTML
@@ -341,8 +384,18 @@ final class Newspack_Popups_Inserter {
 		if ( $previewed_popup_id ) {
 			$found_popup = Newspack_Popups_Model::retrieve_preview_popup( $previewed_popup_id );
 		} elseif ( isset( $atts['id'] ) ) {
-			$found_popup = Newspack_Popups_Model::retrieve_popup_by_id( $atts['id'] );
+			$found_popup = Newspack_Popups_Model::retrieve_popup_by_id( $atts['id'], ! empty( Newspack_Popups_View_As::viewing_as_spec() ) );
 		}
+		if (
+			! $found_popup ||
+			// Bail if it's a non-preview campaign which should not be displayed.
+			( ! self::should_display( $found_popup, true ) && ! Newspack_Popups::previewed_popup_id() ) ||
+			// Only inline popups can be inserted via the  shortcode.
+			! Newspack_Popups_Model::is_inline( $found_popup )
+		) {
+			return;
+		}
+
 		// Wrapping the inline popup in an aside element prevents the markup from being mangled
 		// if the shortcode is the first block.
 		return '<aside>' . Newspack_Popups_Model::generate_popup( $found_popup ) . '</aside>';
@@ -356,7 +409,7 @@ final class Newspack_Popups_Inserter {
 	public static function create_single_popup_access_payload( $popup ) {
 		$popup_id_string = Newspack_Popups_Model::canonize_popup_id( esc_attr( $popup['id'] ) );
 		$frequency       = $popup['options']['frequency'];
-		if ( 'inline' !== $popup['options']['placement'] && 'always' === $frequency ) {
+		if ( Newspack_Popups_Model::is_overlay( $popup ) && 'always' === $frequency ) {
 			$frequency = 'once';
 		}
 		return [
@@ -380,8 +433,31 @@ final class Newspack_Popups_Inserter {
 		if ( ! Newspack_Popups_Segmentation::is_tracking() ) {
 			return;
 		}
+		$shortcoded_popup_ids = array_unique(
+			array_merge(
+				self::get_shortcoded_popups_ids( get_the_content() ),
+				self::get_all_widget_shortcoded_popups_ids()
+			)
+		);
+		$shortcoded_popups    = array_reduce(
+			$shortcoded_popup_ids,
+			function ( $acc, $id ) {
+				$popup_post = get_post( $id );
+				if ( $popup_post ) {
+					$popup_object = Newspack_Popups_Model::create_popup_object( $popup_post );
+					if ( $popup_object ) {
+						$acc[] = $popup_object;
+					}
+				}
+				return $acc;
+			},
+			[]
+		);
 
-		$popups = self::popups_for_post();
+		$popups = array_merge(
+			self::popups_for_post(),
+			$shortcoded_popups
+		);
 		// "Escape hatch" if there's a need to block adding amp-access for pages that have no campaigns.
 		if ( apply_filters( 'newspack_popups_suppress_insert_amp_access', false, $popups ) ) {
 			return;
@@ -431,6 +507,10 @@ final class Newspack_Popups_Inserter {
 				'is_post'    => is_single(),
 			]
 		);
+		$view_as_spec                             = Newspack_Popups_View_As::viewing_as_spec();
+		if ( $view_as_spec ) {
+			$popups_access_provider['authorization'] .= '&view_as=' . wp_json_encode( $view_as_spec );
+		}
 		?>
 		<script id="amp-access" type="application/json">
 			<?php echo wp_json_encode( $popups_access_provider ); ?>
@@ -482,6 +562,33 @@ final class Newspack_Popups_Inserter {
 	}
 
 	/**
+	 * Look for popup shortcodes in a string and return their IDs.
+	 *
+	 * @param string $string String to assess.
+	 * @return array Found shortcoded popups IDs.
+	 */
+	public static function get_shortcoded_popups_ids( $string ) {
+		preg_match_all( '/\[newspack-popup .*\]/', $string, $popup_shortcodes_in_content );
+		if ( empty( $popup_shortcodes_in_content ) ) {
+			return [];
+		} else {
+			return array_unique(
+				array_map(
+					function ( $item ) {
+						preg_match( '/id=["|\'](\d*)/', $item, $matches );
+						if ( empty( $matches ) ) {
+							return null;
+						} else {
+							return $matches[1];
+						}
+					},
+					$popup_shortcodes_in_content[0]
+				)
+			);
+		}
+	}
+
+	/**
 	 * Some popups can only appear on Posts.
 	 *
 	 * @param object $popup The popup to assess.
@@ -498,19 +605,6 @@ final class Newspack_Popups_Inserter {
 	}
 
 	/**
-	 * If Pop-up Frequency is "Test Mode," assess whether it should be shown.
-	 *
-	 * @param object $popup The popup to assess.
-	 * @return bool Should popup be shown based on Test Mode assessment.
-	 */
-	public static function assess_test_mode( $popup ) {
-		if ( 'test' === $popup['options']['frequency'] ) {
-			return is_user_logged_in() && ( current_user_can( 'edit_others_pages' ) || Newspack_Popups::previewed_popup_id() );
-		}
-		return true;
-	}
-
-	/**
 	 * If Pop-up has categories, it should only be shown on posts/pages with those.
 	 *
 	 * @param object $popup The popup to assess.
@@ -519,6 +613,15 @@ final class Newspack_Popups_Inserter {
 	public static function assess_categories_filter( $popup ) {
 		$post_categories  = get_the_category();
 		$popup_categories = get_the_category( $popup['id'] );
+
+		// Filter out "Uncategorized" category which is automatically added to uncategorized posts on publish.
+		$popup_categories = array_filter(
+			$popup_categories,
+			function( $popup_category ) {
+				return 'uncategorized' !== $popup_category->slug;
+			}
+		);
+
 		if ( $post_categories && count( $post_categories ) && $popup_categories && count( $popup_categories ) ) {
 			return array_intersect(
 				array_column( $post_categories, 'term_id' ),
@@ -550,22 +653,78 @@ final class Newspack_Popups_Inserter {
 	 * Should Popup be rendered, based on universal conditions.
 	 *
 	 * @param object $popup The popup to assess.
+	 * @param bool   $skip_context_checks Skip checking context, like if the popup is rendered in a post, and if category/tags are matching.
 	 * @return bool Should popup be shown.
 	 */
-	public static function should_display( $popup ) {
-		// Hide non-test mode campaigns for logged-in users.
-		if ( is_user_logged_in() && 'test' !== $popup['options']['frequency'] ) {
+	public static function should_display( $popup, $skip_context_checks = false ) {
+		if ( 'manual' === $popup['options']['frequency'] ) {
+			return true;
+		}
+
+		$general_conditions = self::assess_is_post( $popup ) &&
+			self::assess_categories_filter( $popup ) &&
+			self::assess_tags_filter( $popup );
+
+		// When using "view as" feature, discard test mode campaigns.
+		if ( Newspack_Popups_View_As::viewing_as_spec() ) {
+			return $general_conditions;
+		}
+		// Hide campaigns for logged-in users.
+		if ( is_user_logged_in() ) {
 			return false;
 		}
 		// Hide overlay campaigns in non-interactive mode, for non-logged-in users.
 		if ( ! is_user_logged_in() && Newspack_Popups_Settings::is_non_interactive() && ! Newspack_Popups_Model::is_inline( $popup ) ) {
 			return false;
 		}
-		return self::assess_is_post( $popup ) &&
-			self::assess_test_mode( $popup ) &&
-			self::assess_categories_filter( $popup ) &&
-			self::assess_tags_filter( $popup ) &&
-			'never' !== $popup['options']['frequency'];
+		if ( $skip_context_checks ) {
+			return true;
+		}
+		return $general_conditions;
+	}
+
+	/**
+	 * When a Text widget is saved and it contains popups shortcode(s), save their IDs as an option.
+	 *
+	 * @param object $instance Widget instance.
+	 * @param object $new_instance New widget instance.
+	 * @param object $old_instance Old widget instance.
+	 * @param object $widget Widget object.
+	 * @return object Widget instance.
+	 */
+	public static function save_widgets_shortcoded_popup_ids( $instance, $new_instance, $old_instance, $widget ) {
+		if ( 'widget_text' === $widget->option_name ) {
+			$value                = get_option( 'newspack_popups_widget_shortcode_popups_ids', [] );
+			$value[ $widget->id ] = self::get_shortcoded_popups_ids( $new_instance['text'] );
+			update_option( 'newspack_popups_widget_shortcode_popups_ids', $value );
+		}
+		return $instance;
+	}
+
+	/**
+	 * Get all widget shortcoded popups IDs.
+	 *
+	 * @return array IDs of popups shortcoded in widgets.
+	 */
+	public static function get_all_widget_shortcoded_popups_ids() {
+		return array_reduce(
+			array_values( get_option( 'newspack_popups_widget_shortcode_popups_ids', [] ) ),
+			function ( $acc, $item ) {
+				return array_merge( $acc, $item );
+			},
+			[]
+		);
+	}
+
+	/**
+	 * Remove widgets shortcoded popup IDs.
+	 *
+	 * @param string $widget_id IDs of a widget.
+	 */
+	public static function remove_widgets_shortcoded_popup_ids( $widget_id ) {
+		$value = get_option( 'newspack_popups_widget_shortcode_popups_ids', [] );
+		unset( $value[ $widget_id ] );
+		update_option( 'newspack_popups_widget_shortcode_popups_ids', $value );
 	}
 }
 $newspack_popups_inserter = new Newspack_Popups_Inserter();