+++ wp-content/plugins/newspack-popups/includes/class-newspack-popups.php
@@ -12,10 +12,11 @@ defined( 'ABSPATH' ) || exit;
  */
 final class Newspack_Popups {
 
-	const NEWSPACK_PLUGINS_CPT             = 'newspack_popups_cpt';
-	const NEWSPACK_POPUPS_SITEWIDE_DEFAULT = 'newspack_popups_sitewide_default';
-
-	const NEWSPACK_POPUP_PREVIEW_QUERY_PARAM = 'newspack_popups_preview_id';
+	const NEWSPACK_POPUPS_CPT                   = 'newspack_popups_cpt';
+	const NEWSPACK_POPUPS_SITEWIDE_DEFAULT      = 'newspack_popups_sitewide_default';
+	const NEWSPACK_POPUPS_TAXONOMY              = 'newspack_popups_taxonomy';
+	const NEWSPACK_POPUPS_ACTIVE_CAMPAIGN_GROUP = 'newspack_popups_active_campaign_group';
+	const NEWSPACK_POPUP_PREVIEW_QUERY_PARAM    = 'newspack_popups_preview_id';
 
 	const LIGHTWEIGHT_API_CONFIG_FILE_PATH_LEGACY = WP_CONTENT_DIR . '/../newspack-popups-config.php';
 	const LIGHTWEIGHT_API_CONFIG_FILE_PATH        = WP_CONTENT_DIR . '/newspack-popups-config.php';
@@ -48,14 +49,13 @@ final class Newspack_Popups {
 		add_action( 'admin_notices', [ __CLASS__, 'api_config_missing_notice' ] );
 		add_action( 'init', [ __CLASS__, 'register_cpt' ] );
 		add_action( 'init', [ __CLASS__, 'register_meta' ] );
+		add_action( 'init', [ __CLASS__, 'register_taxonomy' ] );
 		add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'enqueue_block_editor_assets' ] );
 		add_filter( 'display_post_states', [ __CLASS__, 'display_post_states' ], 10, 2 );
 		add_action( 'rest_api_init', [ __CLASS__, 'rest_api_init' ] );
-		add_action( 'save_post_' . self::NEWSPACK_PLUGINS_CPT, [ __CLASS__, 'popup_default_fields' ], 10, 3 );
+		add_action( 'save_post_' . self::NEWSPACK_POPUPS_CPT, [ __CLASS__, 'popup_default_fields' ], 10, 3 );
 
-		if ( filter_input( INPUT_GET, 'newspack_popups_preview_id', FILTER_SANITIZE_STRING ) ) {
-			add_filter( 'show_admin_bar', [ __CLASS__, 'hide_admin_bar_for_preview' ], 10, 2 ); // phpcs:ignore WordPressVIPMinimum.UserExperience.AdminBarRemoval.RemovalDetected
-		}
+		add_filter( 'show_admin_bar', [ __CLASS__, 'show_admin_bar' ], 10, 2 ); // phpcs:ignore WordPressVIPMinimum.UserExperience.AdminBarRemoval.RemovalDetected
 
 		include_once dirname( __FILE__ ) . '/class-newspack-popups-model.php';
 		include_once dirname( __FILE__ ) . '/class-newspack-popups-inserter.php';
@@ -64,6 +64,7 @@ final class Newspack_Popups {
 		include_once dirname( __FILE__ ) . '/class-newspack-popups-segmentation.php';
 		include_once dirname( __FILE__ ) . '/class-newspack-popups-parse-logs.php';
 		include_once dirname( __FILE__ ) . '/class-newspack-popups-donations.php';
+		include_once dirname( __FILE__ ) . '/class-newspack-popups-view-as.php';
 	}
 
 	/**
@@ -96,7 +97,7 @@ final class Newspack_Popups {
 			'taxonomies'   => [ 'category', 'post_tag' ],
 			'menu_icon'    => 'data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyNCIgaGVpZ2h0PSIyNCIgdmlld0JveD0iMCAwIDI0IDI0Ij48cGF0aCBmaWxsPSIjYTBhNWFhIiBkPSJNMTEuOTkgMTguNTRsLTcuMzctNS43M0wzIDE0LjA3bDkgNyA5LTctMS42My0xLjI3LTcuMzggNS43NHpNMTIgMTZsNy4zNi01LjczTDIxIDlsLTktNy05IDcgMS42MyAxLjI3TDEyIDE2eiIvPjwvc3ZnPgo=',
 		];
-		\register_post_type( self::NEWSPACK_PLUGINS_CPT, $cpt_args );
+		\register_post_type( self::NEWSPACK_POPUPS_CPT, $cpt_args );
 	}
 
 	/**
@@ -107,7 +108,7 @@ final class Newspack_Popups {
 			'post',
 			'trigger_type',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -118,7 +119,7 @@ final class Newspack_Popups {
 			'post',
 			'trigger_scroll_progress',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'integer',
 				'single'         => true,
@@ -129,7 +130,7 @@ final class Newspack_Popups {
 			'post',
 			'trigger_delay',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'integer',
 				'single'         => true,
@@ -141,7 +142,7 @@ final class Newspack_Popups {
 			'post',
 			'frequency',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -153,7 +154,7 @@ final class Newspack_Popups {
 			'post',
 			'placement',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -165,7 +166,7 @@ final class Newspack_Popups {
 			'post',
 			'utm_suppression',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -177,7 +178,7 @@ final class Newspack_Popups {
 			'post',
 			'background_color',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -189,7 +190,7 @@ final class Newspack_Popups {
 			'post',
 			'overlay_color',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -201,7 +202,7 @@ final class Newspack_Popups {
 			'post',
 			'overlay_opacity',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'integer',
 				'single'         => true,
@@ -213,7 +214,7 @@ final class Newspack_Popups {
 			'post',
 			'dismiss_text',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -225,7 +226,7 @@ final class Newspack_Popups {
 			'post',
 			'dismiss_text_alignment',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -237,7 +238,7 @@ final class Newspack_Popups {
 			'post',
 			'display_title',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'boolean',
 				'single'         => true,
@@ -249,7 +250,7 @@ final class Newspack_Popups {
 			'post',
 			'selected_segment_id',
 			[
-				'object_subtype' => self::NEWSPACK_PLUGINS_CPT,
+				'object_subtype' => self::NEWSPACK_POPUPS_CPT,
 				'show_in_rest'   => true,
 				'type'           => 'string',
 				'single'         => true,
@@ -268,6 +269,31 @@ final class Newspack_Popups {
 				'auth_callback' => '__return_true',
 			]
 		);
+	}
+
+	/**
+	 * Register Campaigns taxonomy.
+	 */
+	public static function register_taxonomy() {
+		$taxonomy_args = [
+			'labels'        => [
+				'name'          => __( 'Campaign Groups', 'newspack-popups' ),
+				'singular_name' => __( 'Campaign Group', 'newspack-popups' ),
+				'add_new_item'  => __( 'Add Campaign Group', 'newspack-popups' ),
+			],
+			'hierarchical'  => true,
+			'public'        => false,
+			'rewrite'       => false, // phpcs:ignore Squiz.PHP.CommentedOutCode.Found [ 'hierarchical' => true, 'slug' => $prefix . '/category' ]
+			'show_in_menu'  => false,
+			'show_in_rest'  => true,
+			'show_tagcloud' => false,
+			'show_ui'       => true,
+		];
+
+		register_taxonomy( self::NEWSPACK_POPUPS_TAXONOMY, [ self::NEWSPACK_POPUPS_CPT ], $taxonomy_args );
+
+		// Better safe than sorry: https://developer.wordpress.org/reference/functions/register_taxonomy/#more-information.
+		register_taxonomy_for_object_type( self::NEWSPACK_POPUPS_TAXONOMY, [ self::NEWSPACK_POPUPS_CPT ] );
 	}
 
 	/**
@@ -304,7 +330,7 @@ final class Newspack_Popups {
 	public static function enqueue_block_editor_assets() {
 		$screen = get_current_screen();
 
-		if ( self::NEWSPACK_PLUGINS_CPT !== $screen->post_type ) {
+		if ( self::NEWSPACK_POPUPS_CPT !== $screen->post_type ) {
 			if ( 'page' !== $screen->post_type || 'post' !== $screen->post_type ) {
 				// Script for global settings.
 				\wp_enqueue_script(
@@ -333,6 +359,7 @@ final class Newspack_Popups {
 			[
 				'preview_post' => self::preview_post_permalink(),
 				'segments'     => Newspack_Popups_Segmentation::get_segments(),
+				'taxonomy'     => self::NEWSPACK_POPUPS_TAXONOMY,
 			]
 		);
 		\wp_enqueue_style(
@@ -350,7 +377,7 @@ final class Newspack_Popups {
 	 * @param WP_Post $post        The current post object.
 	 */
 	public static function display_post_states( $post_states, $post ) {
-		if ( self::NEWSPACK_PLUGINS_CPT !== $post->post_type ) {
+		if ( self::NEWSPACK_POPUPS_CPT !== $post->post_type ) {
 			return $post_states;
 		}
 		$post_status_object = get_post_status_object( $post->post_status );
@@ -364,33 +391,31 @@ final class Newspack_Popups {
 	}
 
 	/**
-	 * Hide admin bar if previewing the popup.
+	 * Should admin bar be shown.
 	 *
-	 * @return boolean Whether admin bar should be hidden
+	 * @return boolean Whether admin bar should be shown.
 	 */
-	public static function hide_admin_bar_for_preview() {
-		return ! self::previewed_popup_id();
+	public static function show_admin_bar() {
+		return ! self::is_preview_request();
+	}
+
+	/**
+	 * Is it a preview request – a single popup preview or using "view as" feature.
+	 *
+	 * @return boolean Whether it's a preview request.
+	 */
+	public static function is_preview_request() {
+		$view_as_spec = Newspack_Popups_View_As::viewing_as_spec();
+		return self::previewed_popup_id() || false !== $view_as_spec;
 	}
 
 	/**
 	 * Get previewed popup id from the URL.
 	 *
-	 * @param string $url URL, if available.
 	 * @return number|null Popup id, if found in the URL
 	 */
-	public static function previewed_popup_id( $url = null ) {
-		if ( $url ) {
-			$query_params = [];
-			$parsed_url   = wp_parse_url( $url );
-			parse_str(
-				isset( $parsed_url['query'] ) ? $parsed_url['query'] : '',
-				$query_params
-			);
-			$param = self::NEWSPACK_POPUP_PREVIEW_QUERY_PARAM;
-			return isset( $query_params[ $param ] ) ? $query_params[ $param ] : false;
-		} else {
-			return filter_input( INPUT_GET, self::NEWSPACK_POPUP_PREVIEW_QUERY_PARAM, FILTER_SANITIZE_STRING );
-		}
+	public static function previewed_popup_id() {
+		return filter_input( INPUT_GET, self::NEWSPACK_POPUP_PREVIEW_QUERY_PARAM, FILTER_SANITIZE_STRING );
 	}
 
 	/**
@@ -398,7 +423,7 @@ final class Newspack_Popups {
 	 */
 	public static function rest_api_init() {
 		register_rest_field(
-			[ self::NEWSPACK_PLUGINS_CPT ],
+			[ self::NEWSPACK_POPUPS_CPT ],
 			'newspack_popups_is_sitewide_default',
 			[
 				'get_callback'    => function( $post ) {
@@ -434,16 +459,56 @@ final class Newspack_Popups {
 		if ( $update ) {
 			return;
 		}
-		$placement = isset( $_GET['placement'] ) && 'inline' === sanitize_text_field( $_GET['placement'] ) ? 'inline' : 'center'; //phpcs:ignore
+		$type      = isset( $_GET['placement'] ) ? sanitize_text_field( $_GET['placement'] ) : null; //phpcs:ignore
+		$frequency = 'daily';
+
+		switch ( $type ) {
+			case 'overlay-center':
+				$placement = 'center';
+				break;
+			case 'overlay-top':
+				$placement = 'top';
+				break;
+			case 'overlay-bottom':
+				$placement = 'bottom';
+				break;
+			case 'above-header':
+				$placement = 'above_header';
+				$frequency = 'always';
+				break;
+			case 'manual':
+				$placement = 'inline';
+				$frequency = 'manual';
+				break;
+			default:
+				$placement = 'inline';
+				$frequency = 'always';
+				break;
+		}
+
+		switch ( $type ) {
+			case 'overlay-center':
+			case 'overlay-top':
+			case 'overlay-bottom':
+				$dismiss_text = self::get_default_dismiss_text();
+				$trigger_type = 'time';
+				break;
+			case 'above-header':
+			case 'manual':
+			default:
+				$dismiss_text = null;
+				$trigger_type = 'scroll';
+				break;
+		}
 
 		update_post_meta( $post_id, 'background_color', '#FFFFFF' );
 		update_post_meta( $post_id, 'display_title', false );
-		update_post_meta( $post_id, 'dismiss_text', self::get_default_dismiss_text() );
-		update_post_meta( $post_id, 'frequency', 'test' );
+		update_post_meta( $post_id, 'dismiss_text', $dismiss_text );
+		update_post_meta( $post_id, 'frequency', $frequency );
 		update_post_meta( $post_id, 'overlay_color', '#000000' );
 		update_post_meta( $post_id, 'overlay_opacity', 30 );
 		update_post_meta( $post_id, 'placement', $placement );
-		update_post_meta( $post_id, 'trigger_type', 'time' );
+		update_post_meta( $post_id, 'trigger_type', $trigger_type );
 		update_post_meta( $post_id, 'trigger_delay', 3 );
 		update_post_meta( $post_id, 'trigger_scroll_progress', 30 );
 		update_post_meta( $post_id, 'utm_suppression', '' );