+++ wp-content/plugins/newspack-popups/CHANGELOG.md
@@ -1,3 +1,49 @@
+# [1.22.0](https://github.com/Automattic/newspack-popups/compare/v1.21.0...v1.22.0) (2021-01-21)
+
+
+### Bug Fixes
+
+* corrected logic for above header display ([#403](https://github.com/Automattic/newspack-popups/issues/403)) ([3ca5d71](https://github.com/Automattic/newspack-popups/commit/3ca5d7115f75f31d733f5498caac78d4acdc1040))
+* handle missing GA id ([f27f84a](https://github.com/Automattic/newspack-popups/commit/f27f84a39785030481a7c07b5190555fb3e01f73))
+* in preview tab, allow previewing without choosing groups ([#399](https://github.com/Automattic/newspack-popups/issues/399)) ([c9de7e9](https://github.com/Automattic/newspack-popups/commit/c9de7e9b1067f8ca70bafdfd3853ab26b72c88d8))
+* install composer dependencies on release job ([#368](https://github.com/Automattic/newspack-popups/issues/368)) ([f814d27](https://github.com/Automattic/newspack-popups/commit/f814d270d701dec2997981b1cf1da3f9adf66d3b))
+* remove dismiss button preview when empty ([#402](https://github.com/Automattic/newspack-popups/issues/402)) ([688d575](https://github.com/Automattic/newspack-popups/commit/688d57572dfec80300bf0906eff766b6c1f72010))
+* remove inline class for above header campaigns ([#401](https://github.com/Automattic/newspack-popups/issues/401)) ([212f39e](https://github.com/Automattic/newspack-popups/commit/212f39ea47ca77cb707926010445a543df0e6f51))
+* remove trailing question mark from cleaned non-AMP URLs ([#376](https://github.com/Automattic/newspack-popups/issues/376)) ([25363ae](https://github.com/Automattic/newspack-popups/commit/25363aed2e61d349bc8eb6831a303fb3a03d70e2))
+* segmentation and frequency conflict ([#383](https://github.com/Automattic/newspack-popups/issues/383)) ([47a5af0](https://github.com/Automattic/newspack-popups/commit/47a5af09ffd941780fc6f55a2e54f4c45246299f)), closes [#379](https://github.com/Automattic/newspack-popups/issues/379)
+* segmentation category affinity fixes ([16a6500](https://github.com/Automattic/newspack-popups/commit/16a65003a50b2e411dcb7995152ea941ac86d0e3))
+* session read count in view-as logic ([#362](https://github.com/Automattic/newspack-popups/issues/362)) ([33b65c7](https://github.com/Automattic/newspack-popups/commit/33b65c7d6cf1d8a379246bfbbc4ce41fdde22554))
+* view-as-segment should always take min count values ([#404](https://github.com/Automattic/newspack-popups/issues/404)) ([e69a97a](https://github.com/Automattic/newspack-popups/commit/e69a97a698f9c76552b5aa70d671fb0dc435ac67))
+
+
+### Features
+
+* deprecate test mode and never frequency ([#390](https://github.com/Automattic/newspack-popups/issues/390)) ([cc04314](https://github.com/Automattic/newspack-popups/commit/cc04314c8986734e658613ef1f5120be6bfa2012))
+* handle category affinity segment ([fb01269](https://github.com/Automattic/newspack-popups/commit/fb012699d7e81d04fbb70b3888de205626f7e112))
+* initialize campaigns by placement ([#374](https://github.com/Automattic/newspack-popups/issues/374)) ([5db8e97](https://github.com/Automattic/newspack-popups/commit/5db8e97f339ac7061b16c76939f35c12a46b71b0))
+* manual placement ([#373](https://github.com/Automattic/newspack-popups/issues/373)) ([45b0c67](https://github.com/Automattic/newspack-popups/commit/45b0c677a52c76a3622305a59eb185e316be636d))
+* support for campaign groups in campaigns wizard ([#363](https://github.com/Automattic/newspack-popups/issues/363)) ([fdb12db](https://github.com/Automattic/newspack-popups/commit/fdb12db88cc8c534d25c7a6db430864c866a209a))
+* **segmentation:** add session read count handling ([b1a76cb](https://github.com/Automattic/newspack-popups/commit/b1a76cb83b4a24a2c0626d1a6810cd21865d2c02))
+* **view-as:** handle groups ([#349](https://github.com/Automattic/newspack-popups/issues/349)) ([563f3ce](https://github.com/Automattic/newspack-popups/commit/563f3cef14959b5ad26fdd7c5d2f204752fe1443))
+* shortcode ([#321](https://github.com/Automattic/newspack-popups/issues/321)) ([6da09cf](https://github.com/Automattic/newspack-popups/commit/6da09cfcc19b6275f3aa5f7f3e5321cc2f52d178))
+* view-as feature ([#345](https://github.com/Automattic/newspack-popups/issues/345)) ([632bec7](https://github.com/Automattic/newspack-popups/commit/632bec7e882f9ed7b99f83d24e689cbad278e131))
+
+# [1.21.0](https://github.com/Automattic/newspack-popups/compare/v1.20.0...v1.21.0) (2020-12-15)
+
+
+### Bug Fixes
+
+* editor colors ([69472f3](https://github.com/Automattic/newspack-popups/commit/69472f342c5a58e7e6064cd87361cdb7d73c3997))
+* insert time-triggered overlay campaigns above header ([#297](https://github.com/Automattic/newspack-popups/issues/297)) ([d2c473b](https://github.com/Automattic/newspack-popups/commit/d2c473bc3f610953183e8abd319edbea1133dbad)), closes [#4](https://github.com/Automattic/newspack-popups/issues/4)
+
+
+### Features
+
+* **editor:** hide post visibility selector ([ce7aa98](https://github.com/Automattic/newspack-popups/commit/ce7aa98a409d7b45a9100d5d3faa75e7c7a7b51b))
+* add above header placement option ([#292](https://github.com/Automattic/newspack-popups/issues/292)) ([06a863e](https://github.com/Automattic/newspack-popups/commit/06a863efcab8c60e46015d0354c34deca77f48c8)), closes [#260](https://github.com/Automattic/newspack-popups/issues/260)
+* make campaign groups a hierarchical taxonomy ([#344](https://github.com/Automattic/newspack-popups/issues/344)) ([da4b26a](https://github.com/Automattic/newspack-popups/commit/da4b26ae3e4b62e03ed7c22cd067e1e6f00383b6))
+* render a preview of the dismiss button in the editor ([#336](https://github.com/Automattic/newspack-popups/issues/336)) ([fa7dc05](https://github.com/Automattic/newspack-popups/commit/fa7dc05649f9fcbf66cc95b51cafc34acb1ad8df))
+
 # [1.20.0](https://github.com/Automattic/newspack-popups/compare/v1.19.0...v1.20.0) (2020-12-08)